refactor: 348 - update paths /widget-title
