+++ src/app/layouts/base-layout/components/partnered/partnered.tsx
@@ -1,5 +1,5 @@
 import { AwsLogo, EpamLogo, GithubLogo, JetBrainsLogo } from '@/shared/icons';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './partnered.scss';
