+++ src/widgets/about-video/ui/about-video.tsx
@@ -1,4 +1,4 @@
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './about-video.scss';
