+++ src/widgets/alumni/ui/alumni.tsx
@@ -2,7 +2,7 @@ import { alumni } from '../constants';
 import { useWindowSize } from '@/shared/hooks/use-window-size';
 import Image from '@/shared/ui/image';
 import { Paragraph } from '@/shared/ui/paragraph';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './alumni.scss';
