+++ src/widgets/mentoring/ui/mentoring.tsx
@@ -1,4 +1,4 @@
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './mentoring.scss';
