+++ src/widgets/pictures/ui/pictures.tsx
@@ -13,7 +13,7 @@ import { FacebookIcon, InstagramIcon } from '@/shared/icons';
 import Image from '@/shared/ui/image';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { SocialMedia, SocialMediaProps } from '@/shared/ui/social-media';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import './pictures.scss';
