+++ src/widgets/required/ui/required.tsx
@@ -2,7 +2,7 @@ import { courseDataMap } from '../required.data';
 import type { CourseName } from '../required.types';
 import { Actions } from '@/shared/ui/actions/actions';
 import { Subtitle } from '@/shared/ui/subtitle';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './required.scss';
