+++ src/widgets/speakers/ui/speakers.tsx
@@ -2,7 +2,7 @@ import image from '@/shared/assets/speakers-wanted.webp';
 import { EmailIcon } from '@/shared/icons';
 import Image from '@/shared/ui/image';
 import { Subtitle } from '@/shared/ui/subtitle';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './speakers.scss';
