+++ src/widgets/study-path/ui/study-path.tsx
@@ -3,7 +3,7 @@ import { Stages } from './stages';
 import { type DataMap } from '@/app/services/data/courses-data.types';
 import { useDataByName } from '@/shared/hooks/use-data-by-name';
 import { Paragraph } from '@/shared/ui/paragraph';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 
 import './study-path.scss';
