+++ src/widgets/trainers/ui/trainers.tsx
@@ -1,5 +1,5 @@
 import { TrainerCard } from './trainers-card/trainer-card';
-import { WidgetTitle } from '@/shared/ui/widget-title/widget-title';
+import { WidgetTitle } from '@/shared/ui/widget-title';
 import { Trainer } from '@/widgets/trainers/trainers.types';
 
 import './trainers.scss';
